requirements.txt
django==5.2.6
djangorestframework==3.16.1
django-split-settings==1.3.2
django-cors-headers==4.7.0
drf-spectacular==0.28.0
redis==6.4.0
psycopg2-binary==2.9.10
psycopg==3.2.9
httpx==0.28.1
twilio==9.8.1
python-telegram-bot==22.4
python-dotenv==1.1.1
uvicorn==0.35.0
celery==5.5.3
python-telegram-bot==22.4
pytest-django==4.11.1
